refactor: 패키지 구조 리팩토링 및 클래스 이동
