+++ src/main/java/com/found404/marketbee/mypage/calendar/dto/CalendarCreateReq.java
@@ -1,4 +1,4 @@
-package com.found404.marketbee.mypage.dto;
+package com.found404.marketbee.mypage.calendar.dto;
 
 import java.time.LocalDate;
