+++ src/main/java/com/found404/marketbee/mypage/calendar/dto/CalendarDeleteResp.java
@@ -1,3 +1,3 @@
-package com.found404.marketbee.mypage.dto;
+package com.found404.marketbee.mypage.calendar.dto;
 
 public record CalendarDeleteResp(String message) {}
